back-end/requirements/base.txt
django
django-cors-headers
face-recognition
opencv-python

langchain
langchain-core
langchain-community
langchain-openai
langgraph
langchain-mcp-adapters
langchain-experimental

fastmcp
pymysql
python-dotenv
requests
pillow
numpy
pandas
